change indices to u32

+++ src/parser.rs
@@ -158,7 +158,7 @@ impl<'e> Env<'e> {
         };
 
         match new_name {
-            Ptr::E(index, ..) => assert_eq!(index, lean_pos),
+            Ptr::E(index, ..) => assert_eq!(index as usize, lean_pos),
             _ => unreachable!()
         }
 
@@ -176,7 +176,7 @@ impl<'e> Env<'e> {
          };
 
         match new_level {
-            Ptr::E(index, ..) => assert_eq!(index, lean_pos),
+            Ptr::E(index, ..) => assert_eq!(index as usize, lean_pos),
             _ => unreachable!()
         }
 
@@ -216,7 +216,7 @@ impl<'e> Env<'e> {
         };
 
         match new_expr {
-            Ptr::E(index, ..) => assert_eq!(index, lean_pos),
+            Ptr::E(index, ..) => assert_eq!(index as usize, lean_pos),
             _ => unreachable!()
         }
